Weapons deal damage to players
- Weapons now deal damage to players
- On collision enter, the script checks checks the colliding object to see if there's a weapon script attached and takes the base damage variable from that script, and then deals that damage to player health
- if damage is below a certain threshold (0) it will delete the game object

+++ Assets/Scripts/Combat/Weapon.cs
@@ -7,6 +7,6 @@ using UnityEngine;
 public abstract class Weapon : MonoBehaviour 
 {
     public float cooldownTime;
-    public int baseDamage;
+    public float baseDamage;
     public virtual void Fire(){}
 }
